Add weight, pulse, sugar level fields; decimal age support; updated print template

+++ src/app/api/patients/[id]/route.ts
@@ -40,7 +40,7 @@ export async function PUT(
       where: { id },
       data: {
         name,
-        age: parseInt(age),
+        age: parseFloat(age),
         gender,
         phone: phone || null,
         address: address || null,
